修第二轮检视的三条: blind 默认值、offset 检查溢出、死 include
1. blind_ 的 C++ 默认值是 0.8 而 config 是 0.35。launch 把 yaml 加载到
   ns="hand_lio_node", 所以只有 roslaunch 才拿得到 0.35 —— 直接
   rosrun hand_lio hand_lio_node 会静默落到 0.8, 同一个节点两种启动方式
   跑在不同的盲区半径上。默认值对齐成 0.35, 注释改成"两处必须同步",
   不再引用上一个 commit 刚推翻的那句"跟随 Elevator-LIO 实测值"。

2. 上一个 commit 新加的 offset 检查本身有洞。max_off + 4 > (int)point_step
   是有符号运算: offset 是 uint32_t, 0x7FFFFFFF 转成 int 仍是正数、过得了
   上面的 off_x<0 守卫, 而 int 下的 +4 溢出(UB, 实测回绕成 -2147483645)
   反而小于 point_step, 于是把越界的 offset 放行, 接着 base + 2147483647
   就是大幅越界读。改在 size_t 上比较。
   (0xFFFFFFFF 那种转成 int 是负数, 原来的 off_x<0 守卫能拦住; 漏的只有
    0x00000000~0x7FFFFFFF 这一段里过大的值。)

3. 上一个 commit 换掉 pcl::toROSMsg 之后 pcl_conversions 就没有调用点了,
   换成实际需要的 <sensor_msgs/PointCloud2.h>。CMake 的依赖没动。

另外量了一下检视里提到的"processFrame 跑在 odom 回调里且持着 odom_mutex_":
每点开销(lower_bound 插值 + slerp + 两次刚体变换)实测 2 万点/帧 2.25ms、
6 万点/帧 5.15ms。目前不改, 理由是 processFrame 只在雷达帧被覆盖时跑
(约 10 次/秒而不是 200 次), odom 队列 2000 条(200Hz 下 10 秒)能吸收停顿,
而 Mid-360 一帧约 2 万点对 100ms 的帧间隔有 40 倍余量。
要留意的是注入虚拟障碍之后: max_points 上限 40000 加上雷达 2 万点就是
6 万点, 5.15ms 已超过 200Hz 单次回调的预算 —— 调大那个上限之前要重新量。

+++ src/HandLioNode.cpp
@@ -9,7 +9,7 @@
 #include <geometry_msgs/TransformStamped.h>
 #include <pcl/point_cloud.h>
 #include <pcl/point_types.h>
-#include <pcl_conversions/pcl_conversions.h>
+#include <sensor_msgs/PointCloud2.h>
 
 namespace hand_lio
 {
@@ -255,14 +255,17 @@ namespace hand_lio
         // width/height 和各字段 offset 全来自消息本身，必须当不可信输入校验：
         // 越界的 offset 会读到下一个点甚至缓冲区外，而 data 短于声明的点数时
         // base 直接跑飞。上面只校验了"有没有 x/y/z"，挡不住这两种。
-        const int max_off = std::max({off_x, off_y, off_z,
-                                      has_normal ? off_nx : 0,
-                                      has_normal ? off_ny : 0,
-                                      has_normal ? off_nz : 0});
-        if (max_off + 4 > static_cast<int>(msg->point_step))
+        // 比较必须在 size_t 上做，不能用 int：offset 是 uint32_t，0x7FFFFFFF 转成
+        // int 还是正数、过得了上面的 off_x<0 守卫，而 int 下的 max_off+4 会溢出
+        // (UB，实测回绕成负数)从而小于 point_step，反倒把越界的 offset 放行。
+        const size_t max_off = static_cast<size_t>(std::max({off_x, off_y, off_z,
+                                                             has_normal ? off_nx : 0,
+                                                             has_normal ? off_ny : 0,
+                                                             has_normal ? off_nz : 0}));
+        if (max_off + 4 > static_cast<size_t>(msg->point_step))
         {
             ROS_WARN_THROTTLE(1.0,
-                              "[hand_lio] virtual obstacle cloud field offset %d+4 exceeds point_step %u, ignore",
+                              "[hand_lio] virtual obstacle cloud field offset %zu+4 exceeds point_step %u, ignore",
                               max_off, msg->point_step);
             return;
         }
